Add redirect to payment methods page and handle error message
